fix: update to clean up FormUserManage component
