+++ sites/partners/src/users/FormUserManage.tsx
@@ -1,5 +1,5 @@
 import React, { useMemo, useContext, useState, useCallback } from "react"
-import { useForm } from "react-hook-form"
+import { FormProvider, useForm } from "react-hook-form"
 import {
   Button,
   t,
@@ -8,7 +8,6 @@ import {
   GridCell,
   ViewItem,
   Field,
-  FieldGroup,
   Select,
   useMutate,
   AppearanceStyleType,
@@ -20,6 +19,7 @@ import {
 } from "@bloom-housing/ui-components"
 import { RoleOption, roleKeys, AuthContext } from "@bloom-housing/shared-helpers"
 import { Listing, User, UserRolesCreate } from "@bloom-housing/backend-core/types"
+import { JurisdictionAndListingSelection } from "./JurisdictionAndListingSelection"
 
 type FormUserManageProps = {
   mode: "add" | "edit"
@@ -82,10 +82,11 @@ const FormUserManage = ({
     }
   }
 
-  // eslint-disable-next-line @typescript-eslint/unbound-method
-  const { register, errors, getValues, trigger, setValue, watch } = useForm<FormUserManageValues>({
+  const methods = useForm<FormUserManageValues>({
     defaultValues,
   })
+  // eslint-disable-next-line @typescript-eslint/unbound-method
+  const { register, errors, getValues, trigger, setValue } = methods
 
   const jurisdictionOptions = useMemo(() => {
     return jurisdictionList.map((juris) => ({
@@ -147,49 +148,6 @@ const FormUserManage = ({
     })
     return jurisdictionalizedListings
   }, [getValues, listings, setValue, jurisdictionList])
-
-  /**
-   * Control listing checkboxes on select/deselect all listings option
-   */
-  const updateAllCheckboxes = (e: React.ChangeEvent<HTMLInputElement>, key: string) => {
-    const incomingListingIds = listingsOptions[key].map((option) => option.id)
-    let currentListingIds = getValues("user_listings")
-
-    if (e.target.checked) {
-      const allListingIds = incomingListingIds.reduce(
-        (accum, curr) => {
-          if (!accum.includes(curr)) {
-            // if we are adding listings, make sure we aren't double adding listings
-            accum.push(curr)
-          } else if (!e.target.checked && accum.includes(curr)) {
-            // if we are removing listings
-            accum = accum.filter((elem) => elem !== curr)
-          }
-          return accum
-        },
-        [...(currentListingIds || [])]
-      )
-      setValue("user_listings", allListingIds)
-    } else {
-      if (currentListingIds && !Array.isArray(currentListingIds)) {
-        currentListingIds = [currentListingIds]
-      } else if (!currentListingIds) {
-        currentListingIds = []
-      }
-      const allListingIds = currentListingIds.filter((elem) => !incomingListingIds.includes(elem))
-
-      setValue("user_listings", allListingIds.length === 0 ? null : allListingIds)
-    }
-  }
-
-  const updateAllJurisdictionCheckboxes = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (!e.target.checked) {
-      setValue("jurisdictions", [])
-    } else {
-      const allJurisdictionIds = jurisdictionOptions.map((option) => option.id)
-      setValue("jurisdictions", allJurisdictionIds)
-    }
-  }
 
   const { mutate: sendInvite, isLoading: isSendInviteLoading } = useMutate()
   const { mutate: resendConfirmation, isLoading: isResendConfirmationLoading } = useMutate()
@@ -344,165 +302,8 @@ const FormUserManage = ({
     )
   }
 
-  const selectedRoles = watch("role")
-  const selectedJurisdictions = watch("jurisdictions")
-
-  let jurisdictionAndListingSelection = null
-
-  if (profile?.roles?.isAdmin) {
-    if (selectedRoles === RoleOption.JurisdictionalAdmin) {
-      jurisdictionAndListingSelection = (
-        <GridSection title={t("t.jurisdiction")} columns={4}>
-          <GridCell>
-            <ViewItem>
-              <Select
-                id="jurisdictions"
-                name="jurisdictions"
-                label={t("t.jurisdiction")}
-                placeholder={t("t.selectOne")}
-                labelClassName="sr-only"
-                register={register}
-                controlClassName="control"
-                keyPrefix="users"
-                options={jurisdictionOptions}
-                error={!!errors?.jurisdictions}
-                errorMessage={t("errors.requiredFieldError")}
-                validation={{ required: true }}
-              />
-            </ViewItem>
-          </GridCell>
-        </GridSection>
-      )
-    } else if (selectedRoles === RoleOption.Partner) {
-      jurisdictionAndListingSelection = (
-        <>
-          <GridSection title={t("t.jurisdiction")} columns={4}>
-            <GridCell>
-              <ViewItem>
-                <Field
-                  id="jurisdiction_all"
-                  name="jurisdiction_all"
-                  label={t("users.allJurisdictions")}
-                  register={register}
-                  type="checkbox"
-                  inputProps={{
-                    onChange: (e: React.ChangeEvent<HTMLInputElement>) =>
-                      updateAllJurisdictionCheckboxes(e),
-                  }}
-                />
-
-                <FieldGroup
-                  name="jurisdictions"
-                  fields={jurisdictionOptions}
-                  type="checkbox"
-                  register={register}
-                  error={!!errors?.jurisdictions}
-                  errorMessage={t("errors.requiredFieldError")}
-                  validation={{ required: true }}
-                  dataTestId={"jurisdictions"}
-                />
-              </ViewItem>
-            </GridCell>
-          </GridSection>
-          {selectedJurisdictions && (
-            <GridSection columns={4}>
-              {Object.keys(listingsOptions).map((key) => {
-                if (!selectedJurisdictions.includes(key)) {
-                  return null
-                }
-                const jurisdictionLabel = jurisdictionOptions.find((elem) => elem.id === key)?.label
-                return (
-                  <GridCell key={`listings_${key}`}>
-                    <GridSection
-                      title={t("users.jurisdictionalizedListings", {
-                        jurisdiction: jurisdictionLabel,
-                      })}
-                      columns={1}
-                    >
-                      <GridCell>
-                        <ViewItem>
-                          <Field
-                            id={`listings_all_${key}`}
-                            name={`listings_all_${key}`}
-                            label={t("users.alljurisdictionalizedListings", {
-                              jurisdiction: jurisdictionLabel,
-                            })}
-                            register={register}
-                            type="checkbox"
-                            inputProps={{
-                              onChange: (e: React.ChangeEvent<HTMLInputElement>) =>
-                                updateAllCheckboxes(e, key),
-                            }}
-                          />
-
-                          <FieldGroup
-                            name="user_listings"
-                            fields={listingsOptions[key]}
-                            type="checkbox"
-                            register={register}
-                            error={!!errors?.user_listings}
-                            errorMessage={t("errors.requiredFieldError")}
-                            validation={{ required: true }}
-                            dataTestId={`listings_${jurisdictionLabel}`}
-                          />
-                        </ViewItem>
-                      </GridCell>
-                    </GridSection>
-                  </GridCell>
-                )
-              })}
-            </GridSection>
-          )}
-        </>
-      )
-    }
-  } else if (profile?.roles?.isJurisdictionalAdmin) {
-    if (selectedRoles === RoleOption.Partner && selectedJurisdictions) {
-      jurisdictionAndListingSelection = (
-        <GridSection title={t("nav.listings")} columns={1}>
-          {Object.keys(listingsOptions).map((key) => {
-            if (!selectedJurisdictions.includes(key)) {
-              return null
-            }
-            const jurisdictionLabel = jurisdictionOptions.find((elem) => elem.id === key)?.label
-            return (
-              <GridCell key={`listings_${key}`}>
-                <ViewItem>
-                  <Field
-                    id={`listings_all_${key}`}
-                    name={`listings_all_${key}`}
-                    label={t("users.alljurisdictionalizedListings", {
-                      jurisdiction: jurisdictionLabel,
-                    })}
-                    register={register}
-                    type="checkbox"
-                    inputProps={{
-                      onChange: (e: React.ChangeEvent<HTMLInputElement>) =>
-                        updateAllCheckboxes(e, key),
-                    }}
-                  />
-
-                  <FieldGroup
-                    name="user_listings"
-                    fields={listingsOptions[key]}
-                    type="checkbox"
-                    register={register}
-                    error={!!errors?.user_listings}
-                    errorMessage={t("errors.requiredFieldError")}
-                    validation={{ required: true }}
-                    dataTestId={`listings_${jurisdictionLabel}`}
-                  />
-                </ViewItem>
-              </GridCell>
-            )
-          })}
-        </GridSection>
-      )
-    }
-  }
-
   return (
-    <>
+    <FormProvider {...methods}>
       <Form onSubmit={() => false}>
         <div className="border rounded-md p-8 bg-white">
           <GridSection
@@ -605,7 +406,10 @@ const FormUserManage = ({
               </ViewItem>
             </GridCell>
           </GridSection>
-          {jurisdictionAndListingSelection}
+          <JurisdictionAndListingSelection
+            jurisdictionOptions={jurisdictionOptions}
+            listingsOptions={listingsOptions}
+          />
         </div>
 
         <div className="mt-6">
@@ -689,7 +493,7 @@ const FormUserManage = ({
       >
         {t("users.doYouWantDeleteUser")}
       </Modal>
-    </>
+    </FormProvider>
   )
 }
